+++ crates/xflags-macros/tests/data/empty.rs
@@ -1,0 +1,4 @@
+xflags! {
+    cmd empty {
+    }
+}